* More fixes to functions with money
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@3502 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ luascript.cpp
@@ -8415,7 +8415,7 @@ int32_t LuaInterface::luaDoPlayerAddStamina(lua_State* L)
 int32_t LuaInterface::luaDoPlayerSetBalance(lua_State* L)
 {
 	//doPlayerSetBalance(cid, balance)
-	uint32_t balance = popNumber(L);
+	uint64_t balance = popNumber(L);
 
 	ScriptEnviroment* env = getEnv();
 	if(Player* player = env->getPlayerByUID(popNumber(L)))
